Lesson/subject relation is now composition-based
Previously, Lessons inherited from a subject. But the hierarchy should
be that Subjects are classes that get created in limited number, and
then Lessons & SubjectUIElements get spawned off of them. If Lesson
inherited from Subject, it would be the case that we'd be creating
Lessons (and those would create somewhat hidden Subjects)